+++ Assets/Project/Scenes/StageList.unity
@@ -216,7 +216,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 5
+      value: Stage Card 5
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -330,7 +330,7 @@ GameObject:
   - component: {fileID: 360445769}
   - component: {fileID: 360445770}
   m_Layer: 0
-  m_Name: QuestList Scene Presenter
+  m_Name: StageList Scene Presenter
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -363,7 +363,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 551b2bc4de2ad87459152751afbe3114, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  questListView: {fileID: 1115367387}
+  stageCardListView: {fileID: 1115367387}
 --- !u!1 &425771786
 GameObject:
   m_ObjectHideFlags: 0
@@ -480,7 +480,7 @@ GameObject:
   - component: {fileID: 830426591}
   - component: {fileID: 830426592}
   m_Layer: 5
-  m_Name: QuestCard List
+  m_Name: StageCard List
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -593,7 +593,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 2
+      value: Stage Card 2
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -719,7 +719,7 @@ GameObject:
   - component: {fileID: 1115367386}
   - component: {fileID: 1115367387}
   m_Layer: 5
-  m_Name: QuestListView
+  m_Name: StageList View
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -905,7 +905,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 4
+      value: Stage Card 4
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -1184,7 +1184,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 6
+      value: Stage Card 6
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -1297,7 +1297,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 7
+      value: Stage Card 7
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -1410,7 +1410,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 3
+      value: Stage Card 3
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
@@ -1523,7 +1523,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 4193384719521454248, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Name
-      value: QuestCard 1
+      value: Stage Card 1
       objectReference: {fileID: 0}
     - target: {fileID: 6356673538044290922, guid: 693b10fab4d13544c9fc36a3d473eef5, type: 3}
       propertyPath: m_Pivot.x
